Move drag-and-drop handling from ToolsTab to MainWindow
Enables dropping .sav files on main window instead of just tools tab. Centralizes drag event handlers and drop overlay in MainWindow for broader file acceptance.
